remove scala, kotline and groovy
