+++ camel-idea-plugin/src/main/java/org/apache/camel/idea/gutter/CamelRouteLineMarkerProvider.java
@@ -103,7 +103,7 @@ public class CamelRouteLineMarkerProvider extends RelatedItemLineMarkerProvider 
 
         Icon icon = getCamelPreferenceService().getCamelIcon();
 
-        if (CamelIdeaUtils.isCamelRouteStartExpression(element)) {
+        if (getCamelIdeaUtils().isCamelRouteStartExpression(element)) {
 
             // evaluate the targets lazy
             NotNullLazyValue<Collection<? extends PsiElement>> targets = new NotNullLazyValue<Collection<? extends PsiElement>>() {
@@ -173,7 +173,7 @@ public class CamelRouteLineMarkerProvider extends RelatedItemLineMarkerProvider 
         if (element == null) {
             return false;
         }
-        return CamelIdeaUtils.isCamelRouteStartExpression(element);
+        return getCamelIdeaUtils().isCamelRouteStartExpression(element);
     }
 
     /**
@@ -316,4 +316,7 @@ public class CamelRouteLineMarkerProvider extends RelatedItemLineMarkerProvider 
         return null;
     }
 
+    private static CamelIdeaUtils getCamelIdeaUtils() {
+        return ServiceManager.getService(CamelIdeaUtils.class);
+    }
 }
