+++ camel-idea-plugin/src/main/java/org/apache/camel/idea/service/extension/idea/YamlIdeaUtils.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.idea.service.extension;
+package org.apache.camel.idea.service.extension.idea;
 
 import java.util.Optional;
